+++ packages/core/src/NavigationBase.ts
@@ -1,4 +1,4 @@
-import { ParamsDeserialiser, ParamsSerialiser } from "./common/types";
+import { ParamsDeserializer, ParamsSerializer } from "./common/types";
 import HistoryBase from "./HistoryBase";
 import MetaData from "./MetaData";
 
@@ -35,8 +35,8 @@ export default abstract class NavigationBase {
     protected abstract _history: HistoryBase;
     protected readonly _disableBrowserRouting: boolean;
     protected _currentParams: {[key:string]: any} = {};
-    private _paramsSerialiser?: ParamsSerialiser;
-    private _paramsDeserialiser?: ParamsDeserialiser;
+    private _paramsSerializer?: ParamsSerializer;
+    private _paramsDeserializer?: ParamsDeserializer;
     protected _dispatchEvent: ((event: Event) => Promise<boolean>) | null = null;
 
     constructor(_routerId: number, _disableBrowserRouting: boolean = false, _defaultRoute: string | null = null) {
@@ -68,17 +68,17 @@ export default abstract class NavigationBase {
     abstract goBack(options?: GoBackOptions): void;
 
     searchParamsToObject(searchPart: string) {
-        const deserialiser = this._paramsDeserialiser || this._history.searchParamsToObject;
-        this._currentParams = deserialiser(searchPart) || {};
+        const deserializer = this._paramsDeserializer || this._history.searchParamsToObject;
+        this._currentParams = deserializer(searchPart) || {};
         return this._currentParams;
     }
 
     searchParamsFromObject(params: {[key: string]: any}) {
         try {
-            const serialiser = this._paramsSerialiser || function(paramsObj) {
+            const serializer = this._paramsSerializer || function(paramsObj) {
                 return new URLSearchParams(paramsObj).toString();
             }
-            return serialiser(params);
+            return serializer(params);
         } catch (e) {
             console.error(e);
             console.warn("Non JSON serialisable value was passed as route param to Anchor.");
@@ -90,20 +90,20 @@ export default abstract class NavigationBase {
         this._dispatchEvent = _dispatchEvent;
     }
 
-    set paramsDeserialiser(_paramsDeserialiser: ParamsDeserialiser | undefined) {
-        this._paramsDeserialiser = _paramsDeserialiser;
+    set paramsDeserializer(_paramsDeserializer: ParamsDeserializer | undefined) {
+        this._paramsDeserializer = _paramsDeserializer;
     }
 
-    set paramsSerialiser(_paramsSerialiser: ParamsSerialiser | undefined) {
-        this._paramsSerialiser = _paramsSerialiser;
+    set paramsSerializer(_paramsSerializer: ParamsSerializer | undefined) {
+        this._paramsSerializer = _paramsSerializer;
     }
 
-    get paramsDeserialiser(): ParamsDeserialiser | undefined {
-        return this._paramsDeserialiser;
+    get paramsDeserializer(): ParamsDeserializer | undefined {
+        return this._paramsDeserializer;
     }
 
-    get paramsSerialiser(): ParamsSerialiser | undefined {
-        return this._paramsSerialiser;
+    get paramsSerializer(): ParamsSerializer | undefined {
+        return this._paramsSerializer;
     }
 
     get history() {